When a Proxied request fails (status == 0), report back a 555 to djb, this way the client actually gets a response.
Failure can be caused by a variety of issues, eg remote domain/IP unreachable, DNS problems etc.

We likely at one point will check which states are passed, as it seems that in the unreachable case it is only 1+4, in the "broken" case 1+2+4. We'll add disjjunct/better errors when that happens